girder_plugin/requirements-dev.txt
pip>=9
tox
twine
virtualenv
flake8
flake8-blind-except
flake8-docstrings>=1.3
coverage
mock
pytest>=3.6
pytest-cov==2.5
pytest-girder>=0.1.0a1
pytest-xdist
pyfaidx
fuzzysearch
typing
